Start gpg-agent in main xinitrc
The ssh-key capabilities of the gpg-agent does not know on which tty the
ssh program is running. The pin-entry is therefore display on a default
tty. Thus using the agent is troublesome if multiple ttys are used.
This commit activate the agent on the main tty as that is where it is
used.