keymaps: restore number ordering and rename dvpse
Restore the order of the numbers in the number row to be the same as on
regular dvorak and qwerty keyboards. This required the percent sign to
be moved from shift-1 (in qwerty) to shift-+ as the numbers was shifted
one position to the left.

Also rename from dvpse to dvoormak as the layout now deviated
sufficiently from the original Programmer Dvorak layout and keeping it
based on dvp doesn't make much sense.

I never felt that I benefited from having the different placement of the
numbers, it only made it very difficult to switch back and forth between
the custom layout and regular qwerty/dvorak.